requirements.txt
pyTelegramBotAPI==4.25.0
python-dotenv==1.0.1
requests==2.32.3
